+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Reversed look-and-say".
 - Tests for "Holey times".
 - Tests for "Consecutive balanced substrings".
+- Tests for "Palindromic decomposition".
 
 ### Fixed
 - Tests groups for "CGFunge interpreter".